Add/fix various fields based on my journals

+++ src/modules/logs/content/log_event_content/fetch_remote_module_event.rs
@@ -11,7 +11,7 @@ use crate::modules::ship::{ShipModule, ShipType};
 #[serde(rename_all = "PascalCase")]
 pub struct FetchRemoteModuleEvent {
     /// The storage slot number of the stored module.
-    pub storage_slot: u8,
+    pub storage_slot: u16,
 
     /// The module being transferred.
     pub stored_item: ShipModule,
